# core/openapi.yaml
openapi: 3.0.3
info:
  title: Invoice Generator - Core Service
  description: Orchestrator microservice for client management and invoice creation.
  version: 1.0.0

servers:
  - url: http://localhost:{port}
    description: Local development server
    variables:
      port:
        default: 3000

paths:
  /:
    get:
      summary: Health check
      operationId: healthCheck
      tags:
        - App
      responses:
        '200':
          description: Service is running
          content:
            text/plain:
              schema:
                type: string
                example: Hello World!

  /client:
    post:
      summary: Create a client
      operationId: createClient
      tags:
        - Client
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Client'
      responses:
        '201':
          description: Client created successfully
        '409':
          description: Client with this email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /invoice:
    post:
      summary: Create an invoice
      description: >
        Orchestrates invoice creation by looking up the client, generating a PDF,
        and sending it via email. Returns 404 if the client is not found.
        Duplicate requests within 40 seconds return the existing invoice with duplicate: true.
      operationId: createInvoice
      tags:
        - Invoice
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvoiceDto'
      responses:
        '200':
          description: Invoice created successfully or duplicate detected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateInvoiceResponse'
        '404':
          description: Client not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /invoice/{id}:
    get:
      summary: Get invoice status
      description: Returns the current status of an invoice by its public key.
      operationId: getInvoiceStatus
      tags:
        - Invoice
      parameters:
        - name: id
          in: path
          required: true
          description: The invoice public key
          schema:
            type: string
      responses:
        '200':
          description: Invoice status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceStatusResponse'
        '404':
          description: Invoice not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /invoice/internal/{id}:
    get:
      summary: Get invoice status by numeric ID (internal)
      description: >
        Internal endpoint used by workers to check invoice status before processing.
        Protected by internal API key.
      operationId: getInvoiceStatusById
      tags:
        - Invoice
      security:
        - InternalApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The invoice numeric database ID
          schema:
            type: integer
      responses:
        '200':
          description: Invoice status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceInternalStatusResponse'
        '401':
          description: Invalid internal API key
        '404':
          description: Invoice not found

    patch:
      summary: Update invoice status (internal)
      description: >
        Internal endpoint used by the pdf-generator and email-sender services
        to report status back to the core. Protected by a separate internal API key.
      operationId: updateInvoiceStatus
      tags:
        - Invoice
      security:
        - InternalApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The invoice numeric database ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateInvoiceStatusDto'
      responses:
        '200':
          description: Status updated successfully
        '401':
          description: Invalid internal API key
        '404':
          description: Invoice not found

  /invoice/cron/expire-stale:
    patch:
      summary: Expire stale invoices (cron)
      description: >
        Marks all invoices with status "pending" that are older than 24 hours
        as "expired". Call this from an external cron service.
        Protected by internal API key.
      operationId: expireStaleInvoices
      tags:
        - Cron
      security:
        - InternalApiKeyAuth: []
      responses:
        '200':
          description: Stale invoices expired
          content:
            application/json:
              schema:
                type: object
                properties:
                  expired:
                    type: integer
                    description: Number of invoices expired
                    example: 3
        '401':
          description: Invalid internal API key

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: API key for authentication. Required for /client and /invoice endpoints.
    InternalApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Internal API key for service-to-service authentication. Required for /invoice/internal and /invoice/cron endpoints.

  schemas:
    Client:
      type: object
      required:
        - email
        - firstName
        - lastName
        - companyEmail
        - companyName
      properties:
        email:
          type: string
          description: Client email address
          example: john@example.com
        firstName:
          type: string
          description: Client first name
          example: John
        lastName:
          type: string
          description: Client last name
          example: Doe
        companyEmail:
          type: string
          description: Client company email
          example: john@company.com
        companyName:
          type: string
          description: Client company name
          example: Acme Corp

    CreateInvoiceDto:
      type: object
      required:
        - email
        - jobs
      properties:
        email:
          type: string
          description: Client email address (used to look up the client)
          example: john@example.com
        jobs:
          type: object
          description: Map of job names to their amounts in dollars (values must be >= 0)
          additionalProperties:
            type: number
          example:
            Web Development: 1500.00
            UI Design: 800.50

    CreateInvoiceResponse:
      type: object
      properties:
        id:
          type: string
          description: Invoice public key
          example: 550e8400-e29b-41d4-a716-446655440000
        status:
          type: string
          description: Invoice status
          enum:
            - pending
          example: pending
        duplicate:
          type: boolean
          description: Present and true when a duplicate request is detected within 40 seconds
          example: false

    InvoiceStatusResponse:
      type: object
      properties:
        id:
          type: string
          description: Invoice public key
          example: 550e8400-e29b-41d4-a716-446655440000
        status:
          type: string
          description: Invoice status
          enum:
            - pending
            - resolved
            - expired
            - closed
          example: pending

    InvoiceInternalStatusResponse:
      type: object
      properties:
        status:
          type: string
          description: Invoice status
          enum:
            - pending
            - resolved
            - expired
            - closed
          example: pending

    UpdateInvoiceStatusDto:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          description: New invoice status
          enum:
            - pending
            - resolved
            - expired
            - closed
          example: resolved

    Error:
      type: object
      properties:
        statusCode:
          type: integer
          description: HTTP status code
          example: 404
        message:
          type: string
          description: Error message
          example: Client not found